refactor(api): move await to Call, asFlow to EventSourceFactory
This is the first step in a series of PRs that see the deprecation (and
subsequent removal) of blocking and callback style access in `:sdk-api`.

The end goal is to have everything be suspending instead.
