+++ sample/src/main/kotlin/com/pexip/sdk/sample/pinchallenge/PinChallengeWorkflow.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2023 Pexip AS
+ * Copyright 2022-2024 Pexip AS
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,6 @@
  */
 package com.pexip.sdk.sample.pinchallenge
 
-import com.pexip.sdk.api.coroutines.await
 import com.pexip.sdk.api.infinity.InfinityService
 import com.pexip.sdk.api.infinity.InvalidPinException
 import com.pexip.sdk.api.infinity.RequestTokenRequest
@@ -36,8 +35,7 @@ import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-class PinChallengeWorkflow @Inject constructor(private val store: SettingsStore) :
-    StatefulWorkflow<PinChallengeProps, PinChallengeState, PinChallengeOutput, PinChallengeRendering>() {
+class PinChallengeWorkflow @Inject constructor(private val store: SettingsStore) : StatefulWorkflow<PinChallengeProps, PinChallengeState, PinChallengeOutput, PinChallengeRendering>() {
 
     override fun initialState(props: PinChallengeProps, snapshot: Snapshot?): PinChallengeState =
         PinChallengeState()
