+++ sdk-api/api/sdk-api.api
@@ -1,4 +1,5 @@
 public abstract interface class com/pexip/sdk/api/Call {
+	public abstract fun await (Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 	public abstract fun cancel ()V
 	public abstract fun enqueue (Lcom/pexip/sdk/api/Callback;)V
 	public abstract fun execute ()Ljava/lang/Object;
@@ -17,6 +18,7 @@ public abstract interface class com/pexip/sdk/api/EventSource {
 }
 
 public abstract interface class com/pexip/sdk/api/EventSourceFactory {
+	public fun asFlow ()Lkotlinx/coroutines/flow/Flow;
 	public abstract fun create (Lcom/pexip/sdk/api/EventSourceListener;)Lcom/pexip/sdk/api/EventSource;
 }
