+++ sdk-conference/src/main/kotlin/com/pexip/sdk/conference/infinity/internal/RefererImpl.kt
@@ -15,7 +15,6 @@
  */
 package com.pexip.sdk.conference.infinity.internal
 
-import com.pexip.sdk.api.coroutines.await
 import com.pexip.sdk.api.infinity.InfinityService
 import com.pexip.sdk.api.infinity.RequestTokenRequest
 import com.pexip.sdk.api.infinity.RequestTokenResponse
